requirements.txt
# ============================================================
# primus-dlrm requirements
#
# Hardware target: AMD Instinct MI350+ (MI355X gfx950)
# Platform: ROCm 7.1, Python 3.12
#
# IMPORTANT: PyTorch + FBGEMM_GPU must be installed from
# the ROCm index BEFORE installing this file. See below.
# ============================================================

# --- Step 1: Install PyTorch + ROCm (run these first) ---
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm7.1
# pip install fbgemm-gpu --index-url https://download.pytorch.org/whl/rocm7.1
# pip install torchrec

# --- Step 2: Install remaining deps (this file) ---
# pip install -r requirements.txt

# Core
datasets==4.5.0
pyarrow==23.0.1
polars==1.38.1
pyyaml==6.0.3
tqdm==4.67.3
tensorboard==2.20.0
numpy==2.3.5

# Dev / testing
pytest==9.0.2
pytest-timeout==2.4.0
